// src/api/model/ticketAgeSummary.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * 5-thirty
 * backend data aggregate for 6 o'clock
 * OpenAPI spec version: 0.0.5
 */

export interface TicketAgeSummary {
  aboveOneMonth: number;
  oneDay: number;
  oneMonth: number;
  oneWeek: number;
  threeDays: number;
  twoWeeks: number;
}